requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
dateparser==1.0.0
greenlet==1.1.0
idna==2.10
psycopg2==2.9.1
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
regex==2021.7.6
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.20
tzlocal==2.1
urllib3==1.26.6
